Create sym_forecast_bow_shock.py
This is the code I've been working on at the end of term 1. It imports plasma and SYM/H data as a pandas dataframe df_params. This data is used in the forecasting model, adapted from Burton et al. (1975), to predict the geomagnetic indices over a geomagnetic storm cycle, with an initial SYM/H value as the initial condition. Cross-correlation is then considered to gain some insights about the accuracy of the forecast, compared with the real-life calculated SYM/H values.