Gaussian fits to peak of cross-correlation data
Through this fit, we observe that the peak is not a sharp point of maximum cross-correlation, but seems to flatten off at its maxima, indicating the prediction model predicts events accurately from just after 0mins delay time (as one would expect due to the delay from the bow shock to the Earth) for around 60mins. This will be an interesting aspect to quantify. Could it possibly lead to a 'window' of possibility for an event?